[BUGFIX] - updated commands to work independet of working directory and removed bug when directory contains "."

+++ bin/stop.cmd
@@ -3,16 +3,45 @@
 GOTO :CMDSCRIPT
 ::CMDLITERAL
 
+
+OLDCWD=$(pwd)
+CWD="$( cd "$( echo "${BASH_SOURCE[0]%/*}" )" && pwd )"
+CWD=$(sed 's/.\{4\}$//' <<< "$CWD")
+cd "$CWD"
+
+if [ ! -f "$(pwd)/.env" ]; then
+    echo "Environment File missing. Rename .env-dist to .env and customize it before starting this project."
+    exit
+fi
+
 docker-compose  -p "${PWD##*/}" -f docker-data/config/docker-compose.yml down
+
+cd "$OLDCWD"
 exit
 
 :CMDSCRIPT
-for %%* in (.) do set CurrDirName=%%~nx*
-call:toLower CurrDirName
-set CurrDirName=%CurrDirName: =%
-set CurrDirName=%CurrDirName:-=%
 
-docker-compose  -p "%CurrDirName%" -f docker-data/config/docker-compose.yml down
+SET OLDCWD=%cd%
+SET CWD=%~dp0
+SET CWD=%CWD:~0,-5%
+cd "%CWD%"
+
+IF NOT EXIST "%cd%\.env" (
+    echo Environment File missing. Rename .env-dist to .env and customize it before starting this project.
+    EXIT /B
+)
+
+set Projectname=%~dp0
+set Projectname=%Projectname:~0,-5%
+for %%* in (%Projectname%) do set Projectname=%%~nx*
+set Projectname=%Projectname: =%
+set Projectname=%Projectname:-=%
+set Projectname=%Projectname:.=%
+
+docker-compose  -p "%Projectname%" -f docker-data/config/docker-compose.yml down
+
+CD "%OLDCWD%"
+
 EXIT /B
 
 :toLower str -- converts uppercase character to lowercase
